Changed the definition and behavior of Debug::DumpStack.
startAs is now an int and allows to also look back.

The method now always prints out len bytes, not (len - startAt) bytes, but it still prints out full DWORDS.

+++ src/Misc/Debug.h
@@ -34,7 +34,7 @@ public:
 	static void LogFileClose(int tag);
 	static void LogFileRemove();
 	static void DumpObj(void const* data, size_t len);
-	static void DumpStack(REGISTERS *R, size_t len, size_t startAt = 0);
+	static void DumpStack(REGISTERS* R, size_t len, int startAt = 0);
 	static void (_cdecl* Log)(const char* pFormat, ...);
 
 	template <typename T>
